// generated/addon/angular/src/app/api-models/cieb-ben-plan-accum-trans-log.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class CiebBenPlanAccumTransLog {
  seqPlanTransId: number;
  svcRefId: string;
  reqGroupNumber: string;
  reqBenOptCode: string;
  reqClmSrcSysCode: string;
  reqAsOfDate: string;
  reqStateCode: string;
  invoOutcomCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}